move structures and enums out of players.h and into other files
