+++ xlive/Blam/Engine/game/players.h
@@ -4,68 +4,8 @@
 
 #include "input/controllers.h"
 #include "memory/data.h"
-#include "objects/damage_reporting.h"
-#include "objects/emblems.h"
+#include "networking/network_game_definitions.h"
 #include "simulation/machine_id.h"
-
-enum e_player_color : int8
-{
-	_player_color_white,
-	_player_color_steel,
-	_player_color_red,
-	_player_color_orange,
-	_player_color_gold,
-	_player_color_olive,
-	_player_color_green,
-	_player_color_sage,
-	_player_color_cyan,
-	_player_color_teal,
-	_player_color_colbat,
-	_player_color_blue,
-	_player_color_violet,
-	_player_color_purple,
-	_player_color_pink,
-	_player_color_crimson,
-	_player_color_brown,
-	_player_color_tan,
-	k_player_color_count
-};
-
-/*
- *	New Player Representations Read Me
- *
- *	adding a new representation to the project requires adding a new value to the enum below
- *  once you have done that add a new function in game_globals.cpp to prepare the s_game_globals_custom_representation_result
- *  the function should load/create all necessary data required to create the custom representation (fp, body, biped, variant, etc)
- *	then add the function into the k_game_globals_custom_representation_function_table constant
- */
-enum e_character_type : int8
-{
-	_character_type_masterchief = 0,
-	_character_type_dervish = 1,
-	_character_type_spartan = 2,
-	_character_type_elite = 3,
-
-	k_player_character_type_count_original,	// Original count of characters that shipped with the game
-
-	// cartographer added characters
-	_character_type_skeleton = 4,
-	_character_type_flood = 5,
-	_character_type_lmao = 6,
-	k_player_character_type_count,
-
-	k_cartographer_custom_representation_count = k_player_character_type_count - k_player_character_type_count_original
-};
-
-enum e_handicap : int8
-{
-	_handicap_none = 0,
-	_handicap_minor = 1,
-	_handicap_moderate = 2,
-	_handicap_severe = 3,
-
-	k_handicap_count
-};
 
 enum e_player_flags : int16
 {
@@ -73,62 +13,6 @@ enum e_player_flags : int16
 	_player_left_game_bit = 1,
 	_player_first_spawn_bit = 3,
 };
-
-struct s_player_profile
-{
-	e_player_color primary_color;
-	e_player_color secondary_color;
-	e_player_color tertiary_color;
-	e_player_color quaternary_color;
-	e_character_type player_character_type;
-	s_emblem_info emblem_info;
-};
-ASSERT_STRUCT_SIZE(s_player_profile, 8);
-
-struct s_player_profile_traits
-{
-	s_player_profile profile;
-	int32 gap_48;
-	int32 gap_4C;
-};
-ASSERT_STRUCT_SIZE(s_player_profile_traits, 16);
-
-// practically the xuid, but aligned to 4 (hence the split in 2 32 bit variables)
-struct s_player_identifier
-{
-	uint32 id_low;
-	uint32 id_high;
-};
-ASSERT_STRUCT_SIZE(s_player_identifier, 8);
-
-struct s_clan_identifiers
-{
-	uint32 ID_1;
-	uint32 ID_2;
-	uint32 ID_3;
-};
-ASSERT_STRUCT_SIZE(s_clan_identifiers, 12);
-
-struct s_player_properties
-{
-	wchar_t player_name[XUSER_NAME_SIZE];
-	int32 spawn_protection_time;
-	int8 gap_24[28];
-
-	s_player_profile_traits profile_traits;
-	wchar_t clan_name[16];
-	s_clan_identifiers clan_identifiers;
-
-	int8 team_index;
-	e_handicap player_handicap_level;
-	int8 player_displayed_skill;
-	int8 player_overall_skill;
-	int8 player_is_griefer;
-	int8 bungie_user_role;
-	int8 achievement_flags;
-	int8 unk2;
-};
-ASSERT_STRUCT_SIZE(s_player_properties, 132);
 
 #pragma pack(push, 1)
 struct s_player
@@ -230,28 +114,6 @@ private:
 	s_player* m_current_player;
 	data_iterator m_data_iterator;
 };
-
-struct s_persistent_weapon_data
-{
-	e_damage_reporting_type damage_reporting_type_0;
-	int8 gap_1;
-	e_damage_reporting_type damage_reporting_type_1;
-	int8 gap_3;
-	e_damage_reporting_type damage_reporting_type_2;
-	int8 gap_5;
-	int16 field_6;
-};
-
-struct s_persistent_campaign_player
-{
-	bool initialized;
-	int8 pad;
-	s_persistent_weapon_data weapon_0;
-	s_persistent_weapon_data weapon_1;
-	s_persistent_weapon_data weapon_2;
-	uint16 grenade_counts_mask;
-};
-ASSERT_STRUCT_SIZE(s_persistent_campaign_player, 28);
 
 struct s_players_globals
 {
