+++ xlive/Blam/Engine/saved_games/player_profile.h
@@ -1,16 +1,17 @@
 #pragma once
 #include "game_variant.h"
-
-#include "game/players.h"
 #include "input/input_abstraction.h"
+#include "objects/damage_reporting.h"
+#include "objects/emblems.h"
 
-#define k_default_profiles_count 1
+/* constants */
 
-struct s_saved_game_profile_variant_info
+enum
 {
-	int8 data[192];
-	s_game_variant variant;
+	k_default_profiles_count = 1
 };
+
+/* enums */
 
 enum e_saved_game_profile_input_preference_flags : uint32
 {
@@ -24,6 +25,103 @@ enum e_saved_game_profile_input_preference_flags : uint32
 	k_saved_game_profile_input_preferences_count
 };
 
+/*
+ *	New Player Representations Read Me
+ *
+ *	adding a new representation to the project requires adding a new value to the enum below
+ *  once you have done that add a new function in game_globals.cpp to prepare the s_game_globals_custom_representation_result
+ *  the function should load/create all necessary data required to create the custom representation (fp, body, biped, variant, etc)
+ *	then add the function into the k_game_globals_custom_representation_function_table constant
+ */
+enum e_character_type : int8
+{
+	_character_type_masterchief = 0,
+	_character_type_dervish = 1,
+	_character_type_spartan = 2,
+	_character_type_elite = 3,
+
+	k_player_character_type_count_original,	// Original count of characters that shipped with the game
+
+	// cartographer added characters
+	_character_type_skeleton = 4,
+	_character_type_flood = 5,
+	_character_type_lmao = 6,
+	k_player_character_type_count,
+
+	k_cartographer_custom_representation_count = k_player_character_type_count - k_player_character_type_count_original
+};
+
+enum e_player_color : int8
+{
+	_player_color_white,
+	_player_color_steel,
+	_player_color_red,
+	_player_color_orange,
+	_player_color_gold,
+	_player_color_olive,
+	_player_color_green,
+	_player_color_sage,
+	_player_color_cyan,
+	_player_color_teal,
+	_player_color_colbat,
+	_player_color_blue,
+	_player_color_violet,
+	_player_color_purple,
+	_player_color_pink,
+	_player_color_crimson,
+	_player_color_brown,
+	_player_color_tan,
+	k_player_color_count
+};
+
+/* structures */
+
+struct s_player_profile
+{
+	e_player_color primary_color;
+	e_player_color secondary_color;
+	e_player_color tertiary_color;
+	e_player_color quaternary_color;
+	e_character_type player_character_type;
+	s_emblem_info emblem_info;
+};
+ASSERT_STRUCT_SIZE(s_player_profile, 8);
+
+struct s_player_profile_traits
+{
+	s_player_profile profile;
+	int32 gap_48;
+	int32 gap_4C;
+};
+ASSERT_STRUCT_SIZE(s_player_profile_traits, 16);
+
+struct s_persistent_weapon_data
+{
+	e_damage_reporting_type damage_reporting_type_0;
+	int8 gap_1;
+	e_damage_reporting_type damage_reporting_type_1;
+	int8 gap_3;
+	e_damage_reporting_type damage_reporting_type_2;
+	int8 gap_5;
+	int16 field_6;
+};
+
+struct s_persistent_campaign_player
+{
+	bool initialized;
+	int8 pad;
+	s_persistent_weapon_data weapon_0;
+	s_persistent_weapon_data weapon_1;
+	s_persistent_weapon_data weapon_2;
+	uint16 grenade_counts_mask;
+};
+ASSERT_STRUCT_SIZE(s_persistent_campaign_player, 28);
+
+struct s_saved_game_profile_variant_info
+{
+	int8 data[192];
+	s_game_variant variant;
+};
 
 struct s_saved_game_profile_input_binds
 {
@@ -59,6 +157,7 @@ struct s_saved_game_player_profile
 };
 ASSERT_STRUCT_SIZE(s_saved_game_player_profile, 4616);
 
+/* prototypes */
 
 void __cdecl saved_game_player_profile_set_default_variant(void* saved_game_variant);
